mock service restart
