+++ cmd/storagenode-updater/main.go
@@ -9,7 +9,6 @@ import (
 	"bytes"
 	"context"
 	"errors"
-	"fmt"
 	"io"
 	"io/ioutil"
 	"log"
@@ -332,35 +331,6 @@ func unpackBinary(ctx context.Context, archive, target string) (err error) {
 	_, err = sync2.Copy(ctx, newExec, zipedExec)
 	if err != nil {
 		return errs.Combine(err, os.Remove(newExec.Name()))
-	}
-	return nil
-}
-
-func restartService(name string) error {
-	switch runtime.GOOS {
-	case "windows":
-		// TODO: combine stdout with err if err
-		restartSvcBatPath := filepath.Join(os.TempDir(), "restartservice.bat")
-		restartSvcBat, err := os.Create(restartSvcBatPath)
-		if err != nil {
-			return err
-		}
-
-		restartStr := fmt.Sprintf("net stop %s && net start %s", name, name)
-		_, err = restartSvcBat.WriteString(restartStr)
-		if err != nil {
-			return err
-		}
-		if err := restartSvcBat.Close(); err != nil {
-			return err
-		}
-
-		out, err := exec.Command(restartSvcBat.Name()).CombinedOutput()
-		if err != nil {
-			return errs.New("%s", string(out))
-		}
-	default:
-		return nil
 	}
 	return nil
 }
