--- cmd/storagenode-updater/unix_test.go
@@ -1,17 +1,0 @@
-// Copyright (C) 2019 Storj Labs, Inc.
-// See LICENSE for copying information.
-
-// +build !windows
-
-package main_test
-
-import (
-	"testing"
-
-	"storj.io/storj/internal/testcontext"
-)
-
-// NB: noop
-func createTestService(ctx *testcontext.Context, t *testing.T, name, binPath string) (cleanup func()) {
-	return func() {}
-}
